amtrak: Implement a simple search
Searcher.search() now opens a browser and executes a search on the
Amtrak website for a hard-coded route & date.

The major hurdle was discovering that Amtrak is able to detect the
Selenium browser and return a 403 for the internal call made when
actually making the search (clicking the Find Trains button). I found a
library that wraps Selenium and makes some small changes that remove
Amtrak's ability to detect it.

Note that this library, undetected_geckodriver, did not work out of the
box on WSL. I forked it and made a very small patch that fixed it. I set
the project dependency to point to my fork.